// devtools-ui/src/api/model/getGraphResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */
import type { ModuleGraphEdge } from "./moduleGraphEdge";
import type { ModuleGraphGlobalProviderGroup } from "./moduleGraphGlobalProviderGroup";
import type { ModuleGraphNode } from "./moduleGraphNode";

export interface GetGraphResponse {
	globalProviderGroups: ModuleGraphGlobalProviderGroup[];
	modules: ModuleGraphNode[];
	edges: ModuleGraphEdge[];
}