// devtools-ui/src/api/model/trace.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */
import type { ConsoleEntry } from "./consoleEntry";
import type { TraceErrorProperty } from "./traceErrorProperty";
import type { TraceRequest } from "./traceRequest";
import type { TraceSpan } from "./traceSpan";
import type { TraceSpanStatus } from "./traceSpanStatus";

export interface Trace {
	id: string;
	method: string;
	path: string;
	url: string;
	/** @nullable */
	statusCode: number | null;
	request: TraceRequest;
	response: unknown;
	/** @nullable */
	error: TraceErrorProperty;
	startedAt: number;
	durationMs: number;
	status: TraceSpanStatus;
	spans: TraceSpan[];
	console: ConsoleEntry[];
}