// devtools-ui/src/api/model/traceSpanStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */

export const TraceSpanStatus = { ok: "ok", error: "error" } as const;
export type TraceSpanStatus =
	(typeof TraceSpanStatus)[keyof typeof TraceSpanStatus];